+++ test/apps/express.js
@@ -8,6 +8,7 @@ require('../../')({
   agentPort: process.env.AGENT_PORT,
   level: 'info',
   tracing: {
+    timeBetweenHealthcheckCalls: 1000,
     enabled: process.env.TRACING_ENABLED === 'true',
     forceTransmissionStartingAt: 1,
     stackTraceLength: process.env.STACK_TRACE_LENGTH != null ? parseInt(process.env.STACK_TRACE_LENGTH, 10) : 10
@@ -15,10 +16,43 @@ require('../../')({
 });
 
 var express = require('express');
+var semver = require('semver');
 var app = express();
+
+var healthcheckFunction = function() {
+  return 'OK!';
+};
+
+if (semver.satisfies(process.versions.node, '>=6.0.0')) {
+  require('admin').configure({
+    plugins: [
+      require('admin-plugin-healthcheck')({
+        checks: {
+          configurable: function() {
+            healthcheckFunction();
+          }
+        }
+      })
+    ]
+  });
+}
 
 app.get('/return-instana-trace-id', function(req, res) {
   res.send(req.get('x-instana-t'));
+});
+
+app.post('/admin/set-to-unhealthy', function(req, res) {
+  healthcheckFunction = function() {
+    throw new Error('Explicit healthcheck failure');
+  };
+  res.send('OK');
+});
+
+app.post('/admin/set-to-healthy', function(req, res) {
+  healthcheckFunction = function() {
+    return 'OK';
+  };
+  res.send('OK');
 });
 
 app.use(function(req, res) {